+++ jdbc/src/main/java/org/springframework/transaction/TransactionExecutor.java
@@ -1,4 +1,4 @@
-package org.springframework.jdbc.core;
+package org.springframework.transaction;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.datasource.DataSourceUtils;
