+++ jdbc/src/main/java/org/springframework/transaction/TransactionalOperation.java
@@ -1,4 +1,4 @@
-package org.springframework.jdbc.core;
+package org.springframework.transaction;
 
 public interface TransactionalOperation {
